# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CommonGrid API",
    "version": "1.0.0",
...
    "contact": {
      "name": "CommonGrid Support",
      "email": "hello@texturehq.com",
      "url": "https://texturehq.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/TextureHQ/commongrid/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://commongrid.info/api/v1",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key obtained from the Developer Dashboard. Prefix with `cg_`. Only Bearer is accepted; `X-API-Key` and other Authorization schemes are not supported."
      }
    },
    "parameters": {
...
      "slug": {
        "name": "slug",
        "in": "path",
        "required": true,
        "description": "Unique URL-safe identifier for the resource",
        "schema": {
          "type": "string"
        }
      },
      "at": {
        "name": "at",
        "in": "query",
        "description": "Point-in-time read. Reconstructs the entity from version history at the latest version with changed_at ≤ this instant. Accepts YYYY-MM-DD (end of that UTC day) or an ISO-8601 timestamp. Response includes `_as_of: { requested, versionNumber, changedAt }`. Returns 404 when no version exists at or before the requested time.",
        "schema": {
          "type": "string"
        },
        "example": "2025-12-01"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Resource not found"
              }
            }
          }
        }
      },
      "RateLimitedError": {
        "type": "object",
        "description": "Returned when either the hourly or burst rate-limit window is exceeded. Header values describe the window that tripped.",
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id",
              "currentTier",
              "currentLimit",
              "retryAfter"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "RATE_LIMITED"
                ],
                "example": "RATE_LIMITED"
              },
              "message": {
                "type": "string",
                "example": "Too many requests. Please slow down."
              },
              "request_id": {
                "type": "string",
                "example": "req_abc123"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "currentTier": {
                "type": "string",
                "enum": [
                  "anonymous",
                  "registered",
                  "bulk",
                  "write"
                ],
                "example": "registered"
              },
              "currentLimit": {
                "type": "integer",
                "description": "Budget of the window that was enforced (hourly or burst)",
                "example": 100
              },
              "retryAfter": {
                "type": "integer",
                "description": "Seconds until X-RateLimit-Reset for the window that tripped",
                "example": 12
              },
              "docs": {
                "type": "string",
                "format": "uri"
              },
              "upgrade": {
                "type": "string",
                "description": "Present on anonymous 429s — points callers at free API key registration"
              }
            }
          }
        }
      },
...
      "Utility": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable internal identifier"
          },
          "slug": {
            "type": "string",
            "description": "URL-safe unique identifier used in API paths"
          },
          "name": {
            "type": "string",
            "description": "Display name"
          },
          "eiaName": {
            "type": "string",
            "nullable": true,
            "description": "Utility name as reported on EIA Form 861"
          },
          "shortName": {
            "type": "string",
            "nullable": true,
            "description": "Short / display name variant"
          },
          "logo": {
            "type": "string",
            "nullable": true,
            "description": "URL to the utility's logo"
          },
          "website": {
            "type": "string",
            "nullable": true,
            "description": "Primary utility website"
          },
          "eiaId": {
            "type": "string",
            "nullable": true,
            "description": "EIA utility identifier"
          },
          "segment": {
            "type": "string",
            "description": "Ownership / structural category of the utility. Sourced from EIA Form 861 filings plus manual review.",
            "enum": [
              "COMMUNITY_CHOICE_AGGREGATOR",
              "DISTRIBUTION_COOPERATIVE",
              "FEDERAL",
              "GENERATION_AND_TRANSMISSION",
              "INVESTOR_OWNED_UTILITY",
              "JOINT_ACTION_AGENCY",
              "MUNICIPAL_UTILITY",
              "POLITICAL_SUBDIVISION",
              "TRANSMISSION_OPERATOR"
            ]
          },
          "status": {
            "type": "string",
            "description": "Operational status of the utility.",
            "enum": [
              "ACQUIRED",
              "ACTIVE",
              "DEFUNCT",
              "MERGED",
              "PENDING"
            ]
          },
          "customerCount": {
            "type": "integer",
            "nullable": true,
            "description": "Number of retail customers served (EIA-861)"
          },
          "peakDemandMw": {
            "type": "number",
            "nullable": true,
            "description": "Summer peak demand in megawatts"
          },
          "winterPeakDemandMw": {
            "type": "number",
            "nullable": true,
            "description": "Winter peak demand in megawatts"
          },
          "totalRevenueDollars": {
            "type": "number",
            "nullable": true,
            "description": "Total annual revenue in US dollars"
          },
          "totalSalesMwh": {
            "type": "number",
            "nullable": true,
            "description": "Total annual retail sales in MWh"
          },
          "baCode": {
            "type": "string",
            "nullable": true,
            "description": "Balancing authority code (EIA-861)"
          },
          "nercRegion": {
            "type": "string",
            "nullable": true,
            "description": "NERC reliability region"
          },
          "hasGeneration": {
            "type": "boolean",
            "nullable": true,
            "description": "Whether the utility owns generation assets"
          },
          "hasTransmission": {
            "type": "boolean",
            "nullable": true,
            "description": "Whether the utility owns transmission assets"
          },
          "hasDistribution": {
            "type": "boolean",
            "nullable": true,
            "description": "Whether the utility owns distribution assets"
          },
          "amiMeterCount": {
            "type": "integer",
            "nullable": true,
            "description": "Number of advanced (AMI) meters deployed"
          },
          "totalMeterCount": {
            "type": "integer",
            "nullable": true,
            "description": "Total meter count (AMI + traditional)"
          },
          "jurisdiction": {
            "type": "string",
            "nullable": true,
            "description": "Comma-separated state codes the utility operates in"
          },
          "isoId": {
            "type": "string",
            "nullable": true,
            "description": "FK to the ISO this utility operates under, if any"
          },
          "rtoId": {
            "type": "string",
            "nullable": true,
            "description": "FK to the RTO this utility operates under, if any"
          },
          "balancingAuthorityId": {
            "type": "string",
            "nullable": true,
            "description": "FK to the primary balancing authority"
          },
          "generationProviderId": {
            "type": "string",
            "nullable": true,
            "description": "FK to the utility that supplies generation"
          },
          "transmissionProviderId": {
            "type": "string",
            "nullable": true,
            "description": "FK to the utility that supplies transmission"
          },
          "parentId": {
            "type": "string",
            "nullable": true,
            "description": "FK to parent utility (holding company)"
          },
          "successorId": {
            "type": "string",
            "nullable": true,
            "description": "FK to successor utility (post-merger)"
          },
          "serviceTerritoryId": {
            "type": "string",
            "nullable": true,
            "description": "FK to the region describing the utility's service territory"
          },
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "description": "Email / web domains associated with this utility (e.g. ['duke-energy.com']). Useful for domain-scoped joins against external datasets (email sign-ups, news mentions, complaint logs). Backfilled from website + NRECA/EIA-861 where available; editable via the community contribution flow."
          },
          "deprecationReason": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true,
            "description": "Upstream data source identifier (e.g. EIA, HIFLD, OSM)"
          },
          "sourceUrl": {
            "type": "string",
            "nullable": true,
            "description": "Canonical source URL (filing PDF, dataset download, etc.)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When this record was first ingested into CommonGrid"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When this record was last modified"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Soft-delete timestamp (null for active records)"
          },
          "version": {
            "type": "integer",
            "description": "Monotonic version counter incremented on every edit"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "segment",
          "status",
          "createdAt",
          "updatedAt",
          "version"
        ]
      },
...
    }
  },
  "paths": {
...
    "/utilities/{slug}": {
      "get": {
        "operationId": "getUtility",
        "summary": "Get utility by slug",
        "tags": [
          "Utilities"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/slug"
          },
          {
            "$ref": "#/components/parameters/at"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "headers": {
              "X-RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Budget of the window that applied for this response (hourly or burst). Examples: 60, 100, 5000."
              },
              "X-RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in that same window."
              },
              "X-RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix timestamp (seconds) when the applied window resets."
              },
              "X-RateLimit-Tier": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "anonymous",
                    "registered",
                    "bulk",
                    "write"
                  ]
                },
                "description": "Rate-limit tier that was applied for this request."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Utility"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Hourly or burst rate-limit window exceeded. `Retry-After` and `X-RateLimit-*` describe the window that tripped.",
            "headers": {
              "X-RateLimit-Limit": {
                "schema": {
                  "type": "integer"
                },
                "description": "Budget of the window that applied for this response (hourly or burst). Examples: 60, 100, 5000."
              },
              "X-RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                },
                "description": "Requests remaining in that same window."
              },
              "X-RateLimit-Reset": {
                "schema": {
                  "type": "integer"
                },
                "description": "Unix timestamp (seconds) when the applied window resets."
              },
              "X-RateLimit-Tier": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "anonymous",
                    "registered",
                    "bulk",
                    "write"
                  ]
                },
                "description": "Rate-limit tier that was applied for this request."
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Seconds to wait before retrying. Matches X-RateLimit-Reset for the window that returned 429 (burst or hourly)."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitedError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}